after quite a bit of time of my sloness I finally wrote the main function for training things.
we track the losses since this will be done in batches and the tokens we've seen.

we run it in epochs and train the model, then we calc the loss, backpropagate it, optimize to update weights.

Then we have the evaluation step to just log info and then finally print a sample text